Branch with initial PID version

+++ includes/Drone.h
@@ -7,7 +7,7 @@
 #include "Timer.h"
 
 typedef void (*motor_init_t)(port_t, tim_t, uint8_t);     
-typedef void (*motor_setSignal_t)(tim_t, channel_t, uint32_t, uint8_t);    
+typedef void (*motor_setSignal_t)(tim_t, channel_t, uint32_t, uint16_t);   
 typedef void (*motor_start_t)(tim_t, channel_t); 
 typedef void (*motor_stop_t)(tim_t, channel_t); 
 typedef struct motor
